+++ src/AsyncHttpClient.cpp
@@ -10,6 +10,8 @@ void AsyncHttpClient::queue_request(
 
   job_queue.Dispatch([p_future, p_callback]() -> void {
     cpr::Response response = p_future.get();
+
+    HttpClient::log_response(response);
     HttpHeaders headers = HttpClient::get_headers(response);
     auto http_response = Red::MakeHandle<HttpResponse>(response.status_code,
                                                        headers, response.text);
@@ -26,6 +28,8 @@ void AsyncHttpClient::get(const HttpCallback& p_callback,
     return;
   }
   cpr::Header request_headers = HttpClient::build_headers(p_headers.value);
+
+  HttpClient::log_request(HttpMethod::GET, p_url, "", request_headers);
   cpr::AsyncResponse response = cpr::GetAsync(
     HttpClient::ssl_options, cpr::Url{p_url.c_str()}, request_headers);
 
@@ -45,6 +49,7 @@ void AsyncHttpClient::post(const HttpCallback& p_callback,
   if (!request_headers.contains("Content-Type")) {
     request_headers["Content-Type"] = "text/plain";
   }
+  HttpClient::log_request(HttpMethod::POST, p_url, p_body, request_headers);
   cpr::AsyncResponse response =
     cpr::PostAsync(HttpClient::ssl_options, cpr::Url{p_url.c_str()},
                    cpr::Body{p_body.c_str()}, request_headers);
@@ -66,6 +71,9 @@ void AsyncHttpClient::post_form(const HttpCallback& p_callback,
     values.emplace_back(pair.key.c_str(), pair.value.c_str());
   }
   cpr::Header request_headers = HttpClient::build_headers(p_headers.value);
+
+  HttpClient::log_request_form(HttpMethod::POST, p_url, p_form,
+                               request_headers);
   cpr::AsyncResponse response =
     cpr::PostAsync(HttpClient::ssl_options, cpr::Url{p_url.c_str()},
                    cpr::Payload{values.begin(), values.end()}, request_headers);
@@ -85,6 +93,7 @@ void AsyncHttpClient::put(const HttpCallback& p_callback,
   if (!request_headers.contains("Content-Type")) {
     request_headers["Content-Type"] = "text/plain";
   }
+  HttpClient::log_request(HttpMethod::PUT, p_url, p_body, request_headers);
   cpr::AsyncResponse response =
     cpr::PutAsync(HttpClient::ssl_options, cpr::Url{p_url.c_str()},
                   cpr::Body{p_body.c_str()}, request_headers);
@@ -106,6 +115,8 @@ void AsyncHttpClient::put_form(const HttpCallback& p_callback,
     values.emplace_back(pair.key.c_str(), pair.value.c_str());
   }
   cpr::Header request_headers = HttpClient::build_headers(p_headers.value);
+
+  HttpClient::log_request_form(HttpMethod::PUT, p_url, p_form, request_headers);
   cpr::AsyncResponse response =
     cpr::PutAsync(HttpClient::ssl_options, cpr::Url{p_url.c_str()},
                   cpr::Payload{values.begin(), values.end()}, request_headers);
@@ -126,6 +137,7 @@ void AsyncHttpClient::patch(const HttpCallback& p_callback,
   if (!request_headers.contains("Content-Type")) {
     request_headers["Content-Type"] = "text/plain";
   }
+  HttpClient::log_request(HttpMethod::PATCH, p_url, p_body, request_headers);
   cpr::AsyncResponse response =
     cpr::PatchAsync(HttpClient::ssl_options, cpr::Url{p_url.c_str()},
                     cpr::Body{p_body.c_str()}, request_headers);
@@ -147,6 +159,9 @@ void AsyncHttpClient::patch_form(const HttpCallback& p_callback,
     values.emplace_back(pair.key.c_str(), pair.value.c_str());
   }
   cpr::Header request_headers = HttpClient::build_headers(p_headers.value);
+
+  HttpClient::log_request_form(HttpMethod::PATCH, p_url, p_form,
+                               request_headers);
   cpr::AsyncResponse response = cpr::PatchAsync(
     HttpClient::ssl_options, cpr::Url{p_url.c_str()},
     cpr::Payload{values.begin(), values.end()}, request_headers);
@@ -162,6 +177,8 @@ void AsyncHttpClient::delete_(const HttpCallback& p_callback,
     return;
   }
   cpr::Header request_headers = HttpClient::build_headers(p_headers.value);
+
+  HttpClient::log_request(HttpMethod::DELETE_, p_url, "", request_headers);
   cpr::AsyncResponse response = cpr::DeleteAsync(
     HttpClient::ssl_options, cpr::Url{p_url.c_str()}, request_headers);