+++ src/HttpMethod.h
@@ -1,0 +1,16 @@
+#ifndef REDHTTPCLIENT_HTTPMETHOD_H
+#define REDHTTPCLIENT_HTTPMETHOD_H
+
+namespace RedHttpClient {
+
+enum class HttpMethod {
+  GET,
+  POST,
+  PUT,
+  PATCH,
+  DELETE_
+};
+
+}
+
+#endif  //REDHTTPCLIENT_HTTPMETHOD_H